src/requirements.txt
mcp>=0.1.0
cassandra-driver>=3.25.0
docker>=6.0.0
aiofiles>=23.0.0
playwright>=1.40.0
boto3>=1.26.0